requirements.txt
docker-py
pyaml
requests
pytest
pytest-runner
tox
